% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_reports_functions.R
\name{meta_data_json}
\alias{meta_data_json}
\title{Load beta/gamma for karyograph}
\usage{
meta_data_json(
  pair,
  out_file,
  coverage,
  jabba,
  vcf,
  svaba_somatic_vcf,
  strelka,
  tumor_type_final,
  disease,
  primary_site,
  inferred_sex,
  karyograph,
  seqnames_loh = c(1:22),
  seqnames_genome_width = c(1:22, "X", "Y"),
  write_json = TRUE,
  overwrite = FALSE,
  return_table = FALSE
)
}
\description{
Load beta/gamma for karyograph
}
